lib/string: change return types to match C standard
The return type of strspn and strcspn is supposed to be a size_t and not
a signed integer.

TEST=Now the openSIL code can be built with the coreboot headers without
needing to add '-Wno-builtin-declaration-mismatch' or
'-Wno-incompatible-library-redeclaration' to the cflags. Before the
build would error out with various 'mismatch in return type of built-in
function' errors.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/83223
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/lib/string.c
@@ -125,9 +125,9 @@ int strncmp(const char *s1, const char *s2, size_t maxlen)
 	return 0;
 }
 
-int strspn(const char *str, const char *spn)
+size_t strspn(const char *str, const char *spn)
 {
-	int ret = 0;
+	size_t ret = 0;
 
 	while (*str != 0) {
 		const char *p;
@@ -140,9 +140,9 @@ int strspn(const char *str, const char *spn)
 	return ret;
 }
 
-int strcspn(const char *str, const char *spn)
+size_t strcspn(const char *str, const char *spn)
 {
-	int ret = 0;
+	size_t ret = 0;
 
 	while (*str != 0) {
 		const char *p;
